Added initial review remarks about the current implementation
This is a work in progress, which will be further detailed and completed
tomorrow. Therefore sections might change and/or be removed.

Currently I would assume a discussion about the document does not make a lot of
sense. I will mail it to the interested parties once it is finished, to begin
discussion about the different remarks and observations made.